+++ static/scripts/fix-redirect.js
@@ -59,10 +59,7 @@ const redirects = [
 		fromIncludes("/docs/platform/ip-policies/"),
 		"/docs/api/resources/ip-policies/",
 	],
-	[
-		fromIncludes("/docs/platform/botusers/"),
-		"/docs/user-management/#bot-users",
-	],
+	[fromIncludes("/docs/platform/botusers/"), "/docs/iam/service-users/"],
 	[fromIncludes("/docs/platform/dashboard/"), "/docs/user-management/#sso"],
 	[
 		fromIncludes("/docs/cloud-edge/modules/webhook/"),
@@ -448,7 +445,10 @@ const redirects = [
 	],
 
 	// /docs/user-management/* -> /docs/iam/*
-	[fromIncludes("/docs/user-management/#bot-users"), "/docs/iam/bot-users/"],
+	[
+		fromIncludes("/docs/user-management/#bot-users"),
+		"/docs/iam/service-users/",
+	],
 	[
 		fromIncludes("/docs/user-management/#sso"),
 		"/docs/iam/users/#dashboard-access",
@@ -1050,6 +1050,7 @@ const redirects = [
 		),
 		"/docs/using-ngrok-with/docker/desktop/#video-demo---using-ngrok-with-docker-desktop",
 	],
+	[fromIncludes("/docs/iam/bot-users"), "/docs/iam/service-users/"],
 ];
 
 // get current href from window