+++ docs/agent/changelog.md
@@ -173,7 +173,7 @@ mTLS to upgrade immediately.
 ### ngrok Agent 3.10.0 - \[2024-05-23\]
 
 - **ACTION MAY BE REQUIRED**: The domain used for Certificate Revocation List (CRL) checks is now `crl.ngrok-agent.com` to align it with the domain used for session connections. This may require changes to your firewall or proxy settings to allow this outbound connection on port 80, or setting `crl_noverify: true` in the agent config file.
-- Added CLI API support for [Service Users](/iam/bot-users/).
+- Added CLI API support for [Service Users](/iam/service-users/).
 
 ### ngrok Agent 3.9.0 - \[2024-04-24\]